Move bridge canister trait into separate crate
