+++ src/bridge-canister/src/lib.rs
@@ -1,0 +1,14 @@
+mod build_data;
+mod canister;
+mod config;
+mod core;
+mod inspect;
+mod log_config;
+mod memory;
+mod scheduler;
+mod signer;
+
+pub use core::BridgeCore;
+
+pub use canister::BridgeCanister;
+pub use inspect::bridge_inspect;
