+++ src/bridge-canister/src/log_config.rs
