+++ src/bridge-canister/src/memory.rs
